Fixed Play variant.
